server/requirements.txt
requests
PyPtt == 0.9.50
Flask == 2.2.2
Flask-Cors == 3.0.10
opencc-python-reimplemented
snownlp
wordcloud
matplotlib
gTTS
